# server-stub/openapi.yaml
openapi: 3.0.3
info:
  contact:
    email: support@onevoice.com
    name: OneVoice Team
  description: AI-powered video dubbing service backend API
  license:
    name: MIT
  title: OneVoice Backend API
  version: 1.0.0
servers:
- description: Production server
  url: https://api.onevoice.com
- description: Staging server
  url: https://staging-api.onevoice.com
- description: Development server
  url: http://localhost:8000
security:
- bearerAuth: []
...
paths:
  /api/auth/login:
    post:
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
          description: Successful login
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Invalid credentials
      security: []
      summary: User login
      tags:
      - Authentication
  /api/auth/signup:
    post:
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SignupRequest"
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
          description: User created successfully
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Invalid input data
      security: []
      summary: User registration
      tags:
      - Authentication
  /api/auth/forgot-password:
    post:
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/_api_auth_forgot_password_post_request"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/_api_auth_forgot_password_post_200_response"
          description: Password reset email sent
      security: []
      summary: Request password reset
      tags:
      - Authentication
  /api/auth/reset-password:
    post:
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/_api_auth_reset_password_post_request"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/_api_auth_forgot_password_post_200_response"
          description: Password reset successful
      security: []
      summary: Reset password with token
      tags:
      - Authentication
...
  /api/billing/subscription:
    get:
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subscription"
          description: Current subscription retrieved
      summary: Get current subscription
      tags:
      - Billing & Subscriptions
    put:
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/_api_billing_subscription_put_request"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subscription"
          description: Subscription updated successfully
      summary: Update subscription
      tags:
      - Billing & Subscriptions
  /api/billing/subscribe:
    post:
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/_api_billing_subscribe_post_request"
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subscription"
          description: Subscription created successfully
      summary: Create new subscription
      tags:
      - Billing & Subscriptions
...
  /api/billing/history:
    get:
      parameters:
      - explode: true
        in: query
        name: page
        required: false
        schema:
          default: 1
          type: integer
        style: form
      - explode: true
        in: query
        name: perPage
        required: false
        schema:
          default: 20
          type: integer
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/_api_billing_history_get_200_response"
          description: Billing history retrieved
      summary: Get billing history
      tags:
      - Billing & Subscriptions
...
  /api/billing/payment-method:
    put:
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/_api_billing_payment_method_put_request"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/_api_auth_forgot_password_post_200_response"
          description: Payment method updated successfully
      summary: Update default payment method
      tags:
      - Billing & Subscriptions
...
  /api/billing/payment-methods/{id}:
    delete:
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/_api_auth_forgot_password_post_200_response"
          description: Payment method deleted successfully
      summary: Delete payment method
      tags:
      - Billing & Subscriptions
...
components:
  schemas:
    User:
      example:
        full_name: full_name
        avatar_url: https://openapi-generator.tech
        updated_at: 2000-01-23T04:56:07.000+00:00
        created_at: 2000-01-23T04:56:07.000+00:00
        id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
        email: email
        username: username
      properties:
        id:
          format: uuid
          title: id
          type: string
        email:
          format: email
          title: email
          type: string
        username:
          title: username
          type: string
        full_name:
          title: full_name
          type: string
        avatar_url:
          format: uri
          title: avatar_url
          type: string
        created_at:
          format: date-time
          title: created_at
          type: string
        updated_at:
          format: date-time
          title: updated_at
          type: string
      title: User
      type: object
    LoginRequest:
      example:
        password: password
        email: email
      properties:
        email:
          format: email
          title: email
          type: string
        password:
          minLength: 8
          title: password
          type: string
      required:
      - email
      - password
      title: LoginRequest
      type: object
    SignupRequest:
      example:
        password: password
        full_name: full_name
        email: email
        username: username
      properties:
        email:
          format: email
          title: email
          type: string
        password:
          minLength: 8
          title: password
          type: string
        username:
          title: username
          type: string
        full_name:
          title: full_name
          type: string
      required:
      - email
      - password
      - username
      title: SignupRequest
      type: object
    AuthResponse:
      example:
        access_token: access_token
        token_type: bearer
        expires_in: 0
        user:
          full_name: full_name
          avatar_url: https://openapi-generator.tech
          updated_at: 2000-01-23T04:56:07.000+00:00
          created_at: 2000-01-23T04:56:07.000+00:00
          id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
          email: email
          username: username
      properties:
        access_token:
          title: access_token
          type: string
        token_type:
          example: bearer
          title: token_type
          type: string
        expires_in:
          title: expires_in
          type: integer
        user:
          $ref: "#/components/schemas/User"
      title: AuthResponse
      type: object
...
    Subscription:
      example:
        cancel_at_period_end: true
        id: id
        current_period_end: 2000-01-23T04:56:07.000+00:00
        plan_id: plan_id
        current_period_start: 2000-01-23T04:56:07.000+00:00
        status: active
      properties:
        id:
          title: id
          type: string
        plan_id:
          title: plan_id
          type: string
        status:
          enum:
          - active
          - cancelled
          - past_due
          - unpaid
          title: status
          type: string
        current_period_start:
          format: date-time
          title: current_period_start
          type: string
        current_period_end:
          format: date-time
          title: current_period_end
          type: string
        cancel_at_period_end:
          title: cancel_at_period_end
          type: boolean
      title: Subscription
      type: object
...
    Error:
      example:
        details: "{}"
        error: error
        message: message
      properties:
        error:
          title: error
          type: string
        message:
          title: message
          type: string
        details:
          title: details
          type: object
      title: Error
      type: object
    _api_auth_forgot_password_post_request:
      properties:
        email:
          format: email
          title: email
          type: string
      required:
      - email
      title: _api_auth_forgot_password_post_request
      type: object
    _api_auth_forgot_password_post_200_response:
      example:
        message: message
      properties:
        message:
          title: message
          type: string
      title: _api_auth_forgot_password_post_200_response
      type: object
    _api_auth_reset_password_post_request:
      properties:
        token:
          title: token
          type: string
        new_password:
          minLength: 8
          title: new_password
          type: string
      required:
      - new_password
      - token
      title: _api_auth_reset_password_post_request
      type: object
...
    _api_billing_subscription_put_request:
      properties:
        plan_id:
          title: plan_id
          type: string
      required:
      - plan_id
      title: _api_billing_subscription_put_request
      type: object
    _api_billing_subscribe_post_request:
      properties:
        plan_id:
          title: plan_id
          type: string
        payment_method_id:
          title: payment_method_id
          type: string
      required:
      - payment_method_id
      - plan_id
      title: _api_billing_subscribe_post_request
      type: object
    _api_billing_history_get_200_response_invoices_inner:
      example:
        amount: 0.8008281904610115
        created_at: 2000-01-23T04:56:07.000+00:00
        currency: currency
        id: id
        status: status
      properties:
        id:
          title: id
          type: string
        amount:
          title: amount
          type: number
        currency:
          title: currency
          type: string
        status:
          title: status
          type: string
        created_at:
          format: date-time
          title: created_at
          type: string
      title: _api_billing_history_get_200_response_invoices_inner
      type: object
    _api_billing_history_get_200_response:
      example:
        pagination: "{}"
        invoices:
        - amount: 0.8008281904610115
          created_at: 2000-01-23T04:56:07.000+00:00
          currency: currency
          id: id
          status: status
        - amount: 0.8008281904610115
          created_at: 2000-01-23T04:56:07.000+00:00
          currency: currency
          id: id
          status: status
      properties:
        invoices:
          items:
            $ref: "#/components/schemas/_api_billing_history_get_200_response_invoices_inner"
          title: invoices
          type: array
        pagination:
          title: pagination
          type: object
      title: _api_billing_history_get_200_response
      type: object
...
    _api_billing_payment_method_put_request:
      properties:
        payment_method_id:
          title: payment_method_id
          type: string
      required:
      - payment_method_id
      title: _api_billing_payment_method_put_request
      type: object
...
  securitySchemes:
    bearerAuth:
      bearerFormat: JWT
      scheme: bearer
      type: http